pyproject.toml
[tool.poetry]
name = "fastapi-draft"
version = "0.1.0"
description = "template"
authors = ["alina1412 <alina1412@mail.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.14.0"
annotated-types = "0.7.0"
anyio = "4.8.0"
astroid = "3.3.8"
async-timeout = "5.0.1"
asyncpg = "0.30.0"
click = "8.1.8"
dill = "0.3.9"
exceptiongroup = "1.2.2"
fastapi = "0.115.6"
greenlet = "3.1.1"
h11 = "0.14.0"
idna = "3.10"
isort = "5.13.2"
mako = "1.3.8"
markupsafe = "3.0.2"
mccabe = "0.7.0"
platformdirs = "4.3.6"
psycopg2-binary = "2.9.10"
pydantic = "2.10.5"
pydantic-core = "2.27.2"
pylint = "3.3.3"
python-dotenv = "1.0.1"
sniffio = "1.3.1"
sqlalchemy = "2.0.37"
starlette = "0.41.3"
tomli = "2.2.1"
tomlkit = "0.13.2"
typing-extensions = "4.12.2"
uvicorn = "0.34.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
line_length = 79
known_local_folder = []

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true